requirements.txt
azure-ai-language-questionanswering==1.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.7
isodate==0.6.1
msal==1.28.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
